BI-13524: Log the redirecting of messages to the invalid topic.
