+++ src/main/java/uk/gov/companieshouse/alphabeticalcompanysearchconsumer/service/InvalidMessageRouter.java
@@ -1,10 +1,15 @@
 package uk.gov.companieshouse.alphabeticalcompanysearchconsumer.service;
 
+import static uk.gov.companieshouse.alphabeticalcompanysearchconsumer.AlphabeticalCompanySearchConsumerApplication.NAMESPACE;
+import static uk.gov.companieshouse.alphabeticalcompanysearchconsumer.logging.LoggingUtils.getLogMap;
+
 import java.util.Map;
 import org.apache.kafka.clients.producer.ProducerInterceptor;
 import org.apache.kafka.clients.producer.ProducerRecord;
 import org.apache.kafka.clients.producer.RecordMetadata;
 import uk.gov.companieshouse.alphabeticalcompanysearchconsumer.util.MessageFlags;
+import uk.gov.companieshouse.logging.Logger;
+import uk.gov.companieshouse.logging.LoggerFactory;
 import uk.gov.companieshouse.stream.ResourceChangedData;
 
 /**
@@ -12,16 +17,29 @@ import uk.gov.companieshouse.stream.ResourceChangedData;
  */
 public class InvalidMessageRouter implements ProducerInterceptor<String, ResourceChangedData> {
 
+    private static final Logger LOGGER = LoggerFactory.getLogger(NAMESPACE);
+
     private MessageFlags messageFlags;
     private String invalidMessageTopic;
 
     @Override
-    public ProducerRecord<String, ResourceChangedData> onSend(ProducerRecord<String, ResourceChangedData> producerRecord) {
+    public ProducerRecord<String, ResourceChangedData> onSend(
+        ProducerRecord<String, ResourceChangedData> producerRecord) {
         if (messageFlags.isRetryable()) {
             messageFlags.destroy();
             return producerRecord;
         } else {
-            return new ProducerRecord<>(this.invalidMessageTopic, producerRecord.key(), producerRecord.value());
+            final var message = producerRecord.value();
+            final var resourceId = message.getResourceId();
+            final var resourceKind = message.getResourceKind();
+            final var resourceUri = message.getResourceUri();
+            LOGGER.error("Encountered non-retryable exception producing message to topic "
+                    + producerRecord.topic() + " for resource ID " + resourceId +
+                    ", resource kind " + resourceKind + ", resource URI " + resourceUri
+                    + ". Redirecting message to invalid topic " + invalidMessageTopic + ".",
+                getLogMap(message));
+            return new ProducerRecord<>(this.invalidMessageTopic, producerRecord.key(),
+                producerRecord.value());
         }
     }
